Refactored POJOs
